activitylogもファイル分割
API_V1_BASEで利用する sql_partは、せっかく封じ込められるのだから、コンスタラクタに渡してStrategyにすることにした（そうしないと、テスト分割時にhookが大変）。
