+++ src/api/activitylog/activitylog.js
@@ -4,15 +4,14 @@
  * encoding=utf-8
  */
 
-var console_output = require("../debugger.js").console_output;
-
 var lib = require("../factory4require.js");
+var API_PARAM = require("./api_param.js").API_PARAM;
+var API_V1_BASE = require("./api_v1_base.js").API_V1_BASE;
 var factoryImpl = { // require()を使う代わりに、new Factory() する。
-    "sql_parts" : new lib.Factory4Require("./sql_lite_db/sql_lite_db.js"),
+    "sql_parts" : new lib.Factory4Require("./sql_db_io/index.js"),
 };
 var _SQL_CONNECTION_CONFIG = require("../sql_config.js");
 factoryImpl[ "CONFIG_SQL" ] = new lib.Factory(_SQL_CONNECTION_CONFIG.CONFIG_SQL);
-factoryImpl[ "SETUP_KEY" ]  = new lib.Factory( _SQL_CONNECTION_CONFIG.SETUP_KEY );
 factoryImpl[ "MAX_USERS"] = new lib.Factory( _SQL_CONNECTION_CONFIG.MAX_USERS );
 factoryImpl[ "MAX_LOGS" ] = new lib.Factory( _SQL_CONNECTION_CONFIG.MAX_LOGS );
 
@@ -22,197 +21,8 @@ exports.factoryImpl = factoryImpl;
 
 
 
-/**
- * Promiseで受けわたす、APIの引数チェックしたい！
- * device_key, type_value, date_start, date_end, max_count
- */
-var API_PARAM = function(init){
-	this.device_key = init.device_key;
-	this.pass_key = init.pass_key;
-	this.type_value = init.type_value;
-	this.date_start = init.date_start;
-	this.date_end   = init.date_end;
-	this.max_count = init.max_count;
-};
-var isDefined = function( self, prop ){
-	if( !self[prop] ){
-		// ここは、正常系では呼ばれないハズなので「console.log()」を直接呼ぶ。
-		console_output( "[API_PARAM]::" + prop + " is NOT defind" );
-	}
-	return self[prop];
-};
-API_PARAM.prototype.getDeviceKey = function(){ return isDefined( this, "device_key"); };
-API_PARAM.prototype.getPassKey = function(){ return isDefined( this, "pass_key"); };
-API_PARAM.prototype.getTypeValue = function(){ return isDefined( this, "type_value"); };
-API_PARAM.prototype.getStartDate = function(){ return isDefined( this, "date_start"); };
-API_PARAM.prototype.getEndDate   = function(){ return isDefined( this, "date_end"); };
-
-API_PARAM.prototype.getMaxCount = function(){ return isDefined( this, "max_count"); };
-API_PARAM.prototype.setMaxCount = function( maxCount ){ this.max_count = maxCount; };
-exports.API_PARAM = API_PARAM;
 
 
-
-
-/**
- * 基本となるSQLのシークエンス。クラス（というか、プロトタイプ）。
- */
-var API_V1_BASE = function( outJsonBuffer ){
-	this._outJsonData = outJsonBuffer ? outJsonBuffer : {};
-};
-API_V1_BASE.prototype.isOwnerValid = function( inputData ){
-	// 接続元の認証Y/Nを検証。
-	var instance = this;
-	var outJsonData = instance._outJsonData;
-	var paramClass = new API_PARAM( inputData );
-
-	return new Promise(function(resolve,reject){ // アロー演算子で統一してもいいんだけどね⇒this問題の回避は。
-		var config = factoryImpl.CONFIG_SQL.getInstance();
-		var isOwnerValid = factoryImpl.sql_parts.getInstance( "isOwnerValid" );
-		var is_onwer_valid_promise = isOwnerValid( 
-			config.database, 
-			paramClass.getDeviceKey(), 
-			paramClass.getPassKey()
-		);
-		is_onwer_valid_promise.then(function( maxCount ){
-			paramClass.setMaxCount( maxCount );
-			resolve( paramClass ); // ⇒次のthen()が呼ばれる。
-		}).catch(function(err){
-			if( err ){
-				outJsonData[ "errer_on_validation" ] = err;
-			}
-			reject({
-				"http_status" : 401 // Unauthorized
-			}); // ⇒次のcatch()が呼ばれる。
-		});
-	});	
-};
-API_V1_BASE.prototype.isAccessRateValied = function(){};
-/**
- * サブクラスでオーバーライドする。
- */
-API_V1_BASE.prototype.requestSql = function( paramClass ){
-	return Promise.resolve();
-	// paramClass =>
-	// API_PARAM.prototype.getDeviceKey = function(){ return isDefined( this, "device_key"); };
-	// API_PARAM.prototype.getTypeValue = function(){ return isDefined( this, "type_value"); };
-	// API_PARAM.prototype.getStartDate = function(){ return isDefined( this, "date_start"); };
-	// API_PARAM.prototype.getEndDate   = function(){ return isDefined( this, "date_end"); };
-	// API_PARAM.prototype.getMaxCount = function(){ return isDefined( this, "max_count"); };
-};
-API_V1_BASE.prototype.closeNormal = function(){
-	var instance = this;
-	var outJsonData = instance._outJsonData;
-
-	return new Promise((resolve,reject)=>{
-		var config = factoryImpl.CONFIG_SQL.getInstance();
-		var closeConnection = factoryImpl.sql_parts.getInstance().closeConnection;
-		closeConnection( config.database ).then(()=>{
-			resolve({
-				"jsonData" : outJsonData,
-				"status" : 200 // OK 【FixMe】直前までの内容に応じて変更する。
-			});
-		});
-	});
-};
-/**
- * @param{Object} err プロパティ{"http_status" : HTTPエラーコード}があれば、ソレをHTTP応答のステータスとする。
- */
-API_V1_BASE.prototype.closeAnomaly = function( err ){
-	var instance = this;
-	var outJsonData = instance._outJsonData;
-
-	return new Promise((resolve,reject)=>{
-		var config = factoryImpl.CONFIG_SQL.getInstance();
-		var closeConnection = factoryImpl.sql_parts.getInstance().closeConnection;
-		var http_status = (err && err.http_status) ? err.http_status : 500;
-
-		closeConnection( config.database ).then(()=>{
-			resolve({
-			"jsonData" : outJsonData,
-			"status" : http_status
-			}); // 異常系処理を終えたので、戻すのは「正常」。
-		});
-	});
-};
-API_V1_BASE.prototype.run = function( inputData ){ // getほげほげObjectFromGetData( queryFromGet );済みを渡す。
-	var instance = this;
-	var outJsonData = instance._outJsonData;
-
-	if( inputData.invalid && inputData.invalid.length > 0 ){
-		outJsonData[ "error_on_format" ] = "GET or POST format is INVAILD.";
-		return Promise.resolve({
-			"jsonData" : outJsonData,
-			"status" : 400 // Bad Request
-		});
-	}
-
-	return new Promise(function(resolve,reject){
-		var createPromiseForSqlConnection = factoryImpl.sql_parts.getInstance().createPromiseForSqlConnection;
-		
-		createPromiseForSqlConnection(
-			factoryImpl.CONFIG_SQL.getInstance()
-		).then(function(){
-			outJsonData["result"] = "sql connection is OK!";
-			resolve();
-		}).catch(function(err){
-			outJsonData[ "errer_on_connection" ] = err;
-			reject({
-				"http_status" : 401 // Unauthorized
-			}); // ⇒次のcatch()が呼ばれる。)			
-		});
-	}).then(function(){
-		return instance.isOwnerValid( inputData ); // ここは、冒頭の引数そのまま渡す。
-	}).then(function( paramClass ){
-		// ToDo：アクセス頻度のガードを入れる。
-		return Promise.resolve( paramClass );
-	}).then( ( paramClass )=>{
-		return instance.requestSql( paramClass );
-	}).then(function(){
-		// ここまですべて正常終了
-		return instance.closeNormal();
-	}).catch(function(err){
-        // どこかでエラーした⇒エラー応答のjson返す。
-		return instance.closeAnomaly( err );
-	});
-};
-exports.API_V1_BASE = API_V1_BASE;
-
-
-exports.api_vi_activitylog_setup = function( queryFromGet, dataFromPost ){
-	var createPromiseForSqlConnection = factoryImpl.sql_parts.getInstance().createPromiseForSqlConnection;
-	var outJsonData = {};
-	var config = factoryImpl.CONFIG_SQL.getInstance()
-	
-	if( dataFromPost.create_key != factoryImpl.SETUP_KEY.getInstance() ){
-		return Promise.resolve({
-			"jsonData" : outJsonData, // 何も入れないまま。
-			"status" : 403 // Forbidden
-		});
-	}
-	return createPromiseForSqlConnection(
-		config
-	).then( ()=>{
-		var setupTable1st = factoryImpl.sql_parts.getInstance().setupTable1st;
-		return setupTable1st( config.database );
-	}).then( (successResultOfTable)=>{
-		outJsonData [ "tables" ] = successResultOfTable;
-		return Promise.resolve(200);
-	}).catch((err)=>{
-		outJsonData [ "setup_err" ] = err;
-		return Promise.resolve(500);
-	}).then(( httpStatus )=>{
-		var closeConnection = factoryImpl.sql_parts.getInstance().closeConnection;
-		return new Promise((resolve,reject)=>{
-			closeConnection( config.database ).then(()=>{
-				resolve({
-					"jsonData" : outJsonData,
-					"status" : httpStatus
-				});
-			});		
-		})
-	});
-}
 exports.api_vi_activitylog_signup = function( queryFromGet, dataFromPost ){
 	var createPromiseForSqlConnection = factoryImpl.sql_parts.getInstance().createPromiseForSqlConnection;
 	var outJsonData = {};
@@ -308,11 +118,13 @@ exports.api_vi_activitylog_signup = function( queryFromGet, dataFromPost ){
 		});
 	});
 };
+
+
 exports.api_vi_activitylog_remove = function( queryFromGet, dataFromPost ){
 	var API_SHOW = function(){
 		// サブクラスのコンスタラクタ
 		this._outJsonData = {};
-		API_V1_BASE.call( this, this._outJsonData ); // 継承元のコンスタラクタを明示的に呼び出す。
+		API_V1_BASE.call( this, factoryImpl.CONFIG_SQL, factoryImpl.sql_parts, this._outJsonData ); // 継承元のコンスタラクタを明示的に呼び出す。
 	};
 	API_SHOW.prototype = Object.create( API_V1_BASE.prototype );
 	API_SHOW.prototype.requestSql = function( paramClass ){
@@ -332,7 +144,7 @@ exports.api_v1_activitylog_show = function( queryFromGet, dataFromPost ){
 	var API_SHOW = function(){
 		// サブクラスのコンスタラクタ
 		this._outJsonData = {};
-		API_V1_BASE.call( this, this._outJsonData ); // 継承元のコンスタラクタを明示的に呼び出す。
+		API_V1_BASE.call( this, factoryImpl.CONFIG_SQL, factoryImpl.sql_parts, this._outJsonData ); // 継承元のコンスタラクタを明示的に呼び出す。
 	};
 	API_SHOW.prototype = Object.create( API_V1_BASE.prototype );
 	API_SHOW.prototype.requestSql = function( paramClass ){
@@ -374,7 +186,7 @@ exports.api_v1_activitylog_show = function( queryFromGet, dataFromPost ){
 exports.api_v1_activitylog_add = function( queryFromGet, dataFromPost ){
 	var API_ADD = function(){
 		this._outJsonData = {};
-		API_V1_BASE.call( this, this._outJsonData ); // 継承元のコンスタラクタを明示的に呼び出す。
+		API_V1_BASE.call( this, factoryImpl.CONFIG_SQL, factoryImpl.sql_parts, this._outJsonData ); // 継承元のコンスタラクタを明示的に呼び出す。
 	};
 	API_ADD.prototype = Object.create( API_V1_BASE.prototype );
 	API_ADD.prototype.requestSql = function( paramClass ){
@@ -435,19 +247,18 @@ exports.api_v1_activitylog_add = function( queryFromGet, dataFromPost ){
  * バッテリーログをSQLから削除API
  */
 exports.api_v1_activitylog_delete = function( queryFromGet, dataFromPost ){
-	var API_ADD = function(){
+	var API_DEL = function(){
 		this._outJsonData = {};
-		API_V1_BASE.call( this, this._outJsonData ); // 継承元のコンスタラクタを明示的に呼び出す。
+		API_V1_BASE.call( this, factoryImpl.CONFIG_SQL, factoryImpl.sql_parts, this._outJsonData ); // 継承元のコンスタラクタを明示的に呼び出す。
 	};
-	API_ADD.prototype = Object.create( API_V1_BASE.prototype );
-	API_ADD.prototype.requestSql = function( paramClass ){
+	API_DEL.prototype = Object.create( API_V1_BASE.prototype );
+	API_DEL.prototype.requestSql = function( paramClass ){
 		
 		// 対象のログデータをSQLへ要求
 		var outJsonData = this._outJsonData;
 		var config = factoryImpl.CONFIG_SQL.getInstance();
 		var deleteActivityLogWhereDeviceKey = factoryImpl.sql_parts.getInstance().deleteActivityLogWhereDeviceKey;
 		var period = {};
-
 		if( paramClass.getStartDate() ){
 			period["start"] = paramClass.getStartDate();
 		}
@@ -478,7 +289,7 @@ exports.api_v1_activitylog_delete = function( queryFromGet, dataFromPost ){
 			});
 		});
 	};
-	var subInstance = new API_ADD();
+	var subInstance = new API_DEL();
 	var getDeleteObjectFromPostData = factoryImpl.sql_parts.getInstance().getDeleteObjectFromPostData;
 	var inputData = getDeleteObjectFromPostData( dataFromPost );
 	return subInstance.run( inputData );
