+++ src/api/activitylog/api_v1_base.js
@@ -1,0 +1,187 @@
+/**
+ * [api_v1_base.js]
+ * 
+ * encoding=utf-8
+ */
+
+var lib = require("../factory4require.js");
+var API_PARAM = require("./api_param.js").API_PARAM;
+var factoryImpl = { // require()を使う代わりに、new Factory() する。
+    "sql_parts" : new lib.Factory4Require("./sql_db_io/index.js"),
+};
+var _SQL_CONNECTION_CONFIG = require("../sql_config.js");
+factoryImpl[ "CONFIG_SQL" ] = new lib.Factory(_SQL_CONNECTION_CONFIG.CONFIG_SQL);
+factoryImpl[ "SETUP_KEY" ]  = new lib.Factory( _SQL_CONNECTION_CONFIG.SETUP_KEY );
+
+
+// UTデバッグ用のHookポイント。運用では外部公開しないメソッドはこっちにまとめる。
+exports.factoryImpl = factoryImpl;
+
+
+ 
+
+/**
+ * 基本となるSQLのシークエンス。クラス（というか、プロトタイプ）。
+ */
+var API_V1_BASE = function( configSql, sqlParts, outJsonBuffer ){
+	this._config_sql = configSql;
+	this._sql_parts = sqlParts;
+	this._outJsonData = outJsonBuffer ? outJsonBuffer : {};
+};
+API_V1_BASE.prototype.isOwnerValid = function( inputData ){
+	// 接続元の認証Y/Nを検証。
+	var instance = this;
+	var outJsonData = instance._outJsonData;
+	var paramClass = new API_PARAM( inputData );
+
+	return new Promise(function(resolve,reject){ // アロー演算子で統一してもいいんだけどね⇒this問題の回避は。
+		var config = instance._config_sql.getInstance();
+		var isOwnerValid = instance._sql_parts.getInstance( "isOwnerValid" );
+		var is_onwer_valid_promise = isOwnerValid( 
+			config.database, 
+			paramClass.getDeviceKey(), 
+			paramClass.getPassKey()
+		);
+		is_onwer_valid_promise.then(function( maxCount ){
+			paramClass.setMaxCount( maxCount );
+			resolve( paramClass ); // ⇒次のthen()が呼ばれる。
+		}).catch(function(err){
+			if( err ){
+				outJsonData[ "errer_on_validation" ] = err;
+			}
+			reject({
+				"http_status" : 401 // Unauthorized
+			}); // ⇒次のcatch()が呼ばれる。
+		});
+	});	
+};
+API_V1_BASE.prototype.isAccessRateValied = function(){};
+/**
+ * サブクラスでオーバーライドする。
+ */
+API_V1_BASE.prototype.requestSql = function( paramClass ){
+	return Promise.resolve();
+	// paramClass =>
+	// API_PARAM.prototype.getDeviceKey = function(){ return isDefined( this, "device_key"); };
+	// API_PARAM.prototype.getTypeValue = function(){ return isDefined( this, "type_value"); };
+	// API_PARAM.prototype.getStartDate = function(){ return isDefined( this, "date_start"); };
+	// API_PARAM.prototype.getEndDate   = function(){ return isDefined( this, "date_end"); };
+	// API_PARAM.prototype.getMaxCount = function(){ return isDefined( this, "max_count"); };
+};
+API_V1_BASE.prototype.closeNormal = function(){
+	var instance = this;
+	var outJsonData = instance._outJsonData;
+
+	return new Promise((resolve,reject)=>{
+		var config = instance._config_sql.getInstance();
+		var closeConnection = instance._sql_parts.getInstance().closeConnection;
+		closeConnection( config.database ).then(()=>{
+			resolve({
+				"jsonData" : outJsonData,
+				"status" : 200 // OK 【FixMe】直前までの内容に応じて変更する。
+			});
+		});
+	});
+};
+/**
+ * @param{Object} err プロパティ{"http_status" : HTTPエラーコード}があれば、ソレをHTTP応答のステータスとする。
+ */
+API_V1_BASE.prototype.closeAnomaly = function( err ){
+	var instance = this;
+	var outJsonData = instance._outJsonData;
+
+	return new Promise((resolve,reject)=>{
+		var config = instance._config_sql.getInstance();
+		var closeConnection = instance._sql_parts.getInstance().closeConnection;
+		var http_status = (err && err.http_status) ? err.http_status : 500;
+
+		closeConnection( config.database ).then(()=>{
+			resolve({
+			"jsonData" : outJsonData,
+			"status" : http_status
+			}); // 異常系処理を終えたので、戻すのは「正常」。
+		});
+	});
+};
+API_V1_BASE.prototype.run = function( inputData ){ // getほげほげObjectFromGetData( queryFromGet );済みを渡す。
+	var instance = this;
+	var outJsonData = instance._outJsonData;
+
+	if( inputData.invalid && inputData.invalid.length > 0 ){
+		outJsonData[ "error_on_format" ] = "GET or POST format is INVAILD.";
+		return Promise.resolve({
+			"jsonData" : outJsonData,
+			"status" : 400 // Bad Request
+		});
+	}
+
+	return new Promise(function(resolve,reject){
+		var createPromiseForSqlConnection = instance._sql_parts.getInstance().createPromiseForSqlConnection;
+		
+		createPromiseForSqlConnection(
+			instance._config_sql.getInstance()
+		).then(function(){
+			outJsonData["result"] = "sql connection is OK!";
+			resolve();
+		}).catch(function(err){
+			outJsonData[ "errer_on_connection" ] = err;
+			reject({
+				"http_status" : 401 // Unauthorized
+			}); // ⇒次のcatch()が呼ばれる。)			
+		});
+	}).then(function(){
+		return instance.isOwnerValid( inputData ); // ここは、冒頭の引数そのまま渡す。
+	}).then(function( paramClass ){
+		// ToDo：アクセス頻度のガードを入れる。
+		return Promise.resolve( paramClass );
+	}).then( ( paramClass )=>{
+		return instance.requestSql( paramClass );
+	}).then(function(){
+		// ここまですべて正常終了
+		return instance.closeNormal();
+	}).catch(function(err){
+        // どこかでエラーした⇒エラー応答のjson返す。
+		return instance.closeAnomaly( err );
+	});
+};
+exports.API_V1_BASE = API_V1_BASE;
+
+
+
+exports.api_vi_activitylog_setup = function( queryFromGet, dataFromPost ){
+    var createPromiseForSqlConnection = factoryImpl.sql_parts.getInstance().createPromiseForSqlConnection;
+	var outJsonData = {};
+	var config = factoryImpl.CONFIG_SQL.getInstance();
+	
+	if( dataFromPost.create_key != factoryImpl.SETUP_KEY.getInstance() ){
+		return Promise.resolve({
+			"jsonData" : outJsonData, // 何も入れないまま。
+			"status" : 403 // Forbidden
+		});
+	}
+	return createPromiseForSqlConnection(
+		config
+	).then( ()=>{
+		var setupTable1st = factoryImpl.sql_parts.getInstance().setupTable1st;
+		return setupTable1st( config.database );
+	}).then( (successResultOfTable)=>{
+		outJsonData [ "tables" ] = successResultOfTable;
+		return Promise.resolve(200);
+	}).catch((err)=>{
+		outJsonData [ "setup_err" ] = err;
+		return Promise.resolve(500);
+	}).then(( httpStatus )=>{
+		var closeConnection = factoryImpl.sql_parts.getInstance().closeConnection;
+		return new Promise((resolve,reject)=>{
+			closeConnection( config.database ).then(()=>{
+				resolve({
+					"jsonData" : outJsonData,
+					"status" : httpStatus
+				});
+			});		
+		});
+	});
+};
+
+
+
